dashboard: remove the call to show_to? from collections view
This is my first attempt with trying to fix some of the performance
issues described in #1175.

While show_to? isn't the problem, I am still hoping that this helps to
reduce complexity.

Unfortunately some methods we use depend on collections and document
sets being iterated (which is the expensive part), so just move that
code to collections_list for now.