Expand scope of update_next_untranscribed_pages
If a work becomes completely transcribed but then later a page within that
work becomes untranscribed, next_untranscribed_page never gets updated.
This causes the work to be missing from some views where works with
untranscribed pages are listed.

This change causes update_next_untranscribed_pages to also be run if
next_untranscribed_page is currently nil. This was tested by completing a
work, subsequently deleting a transcription for a page, saving, then
verifying in the database that the column went from null to the correct
id.

Signed-off-by: Jacob Creedon <jcreedon@gmail.com>