Update user_params filter to use real_name
This fixes a bug where users were not able to update their display name.
The user_params permit list had display_name, however, per the model that
is an auto-generted field and real_name is the one used in the form.

This has been tested and verified to work correctly now.

Signed-off-by: Jacob Creedon <jcreedon@gmail.com>