Remove display_name from admin edit_user form
The display_name field is intended to be auto generated and leads to
unexpected behavior for the user when it is modified. This removes the
display_name form field from the edit_user form and the user_params permit
list. This also adjusts the real_name form field to have the top-of-form
styling that display_name previously had.

Signed-off-by: Jacob Creedon <jcreedon@gmail.com>